Implementation of Role CRUD operations with unit tests and global exception handling, along with documentation updates in the README.md, AGENTS.md, and docs/api/parkio-api-v1.md files.

+++ src/main/java/com/kasaca/parkio/cajon/repository/CajonRepository.java
@@ -3,5 +3,5 @@ package com.kasaca.parkio.cajon.repository;
 import com.kasaca.parkio.cajon.entity.Cajon;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CajonRepository extends JpaRepository<Cajon,Integer> {
+public interface CajonRepository extends JpaRepository<Cajon, Long> {
 }
